made modal a component
