+++ site/activity-react/components/modal/style.scss
@@ -1,0 +1,260 @@
+.file-page {
+    .file-con {
+        margin-top: 40px;
+        padding: 20px 30px 40px 30px;
+    }
+
+    .file-dir {
+        margin-bottom: 15px;
+
+        .blue {
+            color:rgb(92, 92, 196);
+            &:hover {
+                text-decoration: underline;
+                cursor: pointer;
+            }
+        }
+    }
+
+    .create-btn {
+        display: inline-block;
+        font-size: 16px;
+        padding: 10px;
+        width: 100px;
+        border: 1px solid #e6e6e6; /*no*/
+        border-radius: 4px;
+        margin-right: 30px;
+        cursor: pointer;
+        &:hover {
+            background-color: #e6e6e646;
+        }
+    }
+
+    .head-info {
+        font-size: 24px;
+        font-weight: bold;
+        margin-bottom: 20px;
+    }
+
+    .file-list {
+        .file-line {
+            display: flex;
+            border-bottom: 1px solid #e6e6e6; /*no*/
+
+            &.files {
+                font-size: 14px;
+                line-height: 50px; 
+            }
+            &.header {
+                font-size: 12px;
+                line-height: 50px;
+                color: #999;
+            }
+        }
+        .folder-name {
+            display: block;
+            border: none;
+            flex-grow: 1;
+            height: 50px;
+            font-size: 14px;
+            line-height: 50px;
+        }
+
+        div {
+            flex-grow: 0;
+            flex-shrink: 0;
+            overflow: hidden;
+            text-overflow:ellipsis;
+            white-space: nowrap;
+        }
+        .name {
+            width: 50%;
+            &:hover {
+                cursor: pointer;
+            }
+        }
+        .size {
+            width: 15%;
+        }
+        .last-modify {
+            width: 15%;
+        }
+        .tools {
+            width: 20%;
+            span {
+                margin-right: 10px; 
+                &:hover {
+                    cursor: pointer;
+                    text-decoration: underline;
+                }
+            }
+           
+        } 
+        .show-all-file {
+            margin-top: 16px;
+        }
+    }
+
+}
+.window{
+    
+    z-index: 1000000;
+    background-color: rgba(0,0,0,0.5);
+    position: fixed;
+    height: 100%;
+    width: 100%;
+    top: 0;
+    left: 0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    .outerBox {
+        background-color: #e6e6e6;
+        margin: auto;
+        padding: 30px;
+        height: 35%;
+        width: 50%;
+        top: 0;
+        left: 0;
+        right: 0;
+        bottom: 0;
+        border: 1px solid #726d6d; /*no*/
+        border-radius: 8px;
+    }
+  
+}
+.move-File {
+    .move-file-con {
+        padding-top: 30px;
+        
+
+    .file-dir {
+        margin-bottom: 15px;
+        margin-top: 15px;
+        .blue {
+            color: rgb(92, 92, 196);
+            &:hover {
+                text-decoration: underline;
+                cursor: pointer;
+            }
+        }
+ 
+    }
+
+    .create-btn {
+        display: inline-block;
+        font-size: 16px;
+        padding: 10px;
+        width: 100px;
+        border: 1px solid #4e4d4d; /*no*/
+        border-radius: 4px;
+        margin-right: 30px;
+        cursor: pointer;
+        &:hover {
+            background-color: #635f5f46;
+        }
+    }
+
+    .head-info {
+        font-size: 24px;
+        font-weight: bold;
+        margin-bottom: 20px;
+    }
+
+    .file-list {
+        .file-line {
+            display: flex;
+            align-items: center;
+            border-bottom: 1px solid #b1acac5d; /*no*/
+
+            &.files {
+                font-size: 14px;
+                line-height: 50px; 
+            }
+            &.header {
+                font-size: 12px;
+                line-height: 50px;
+                color: #999;
+            }
+        }
+        .folder-name {
+            display: block;
+            border: none;
+            flex-grow: 1;
+            height: 50px;
+            font-size: 14px;
+            line-height: 50px;
+        }
+
+        div {
+            flex-grow: 0;
+            flex-shrink: 0;
+            overflow: hidden;
+            text-overflow:ellipsis;
+            white-space: nowrap;
+        }
+        .name {
+            width: 50%;
+             &:hover {
+                cursor: pointer;
+            }
+        }
+        .size {
+            width: 15%;
+        }
+        .last-modify {
+            width: 15%;
+        }
+        .tools {
+            width: 20%;
+            span {
+                margin-right: 10px; 
+                &:hover {
+                    cursor: pointer;
+                    text-decoration: underline;
+                }
+            }
+        } 
+        .show-all-file {
+            margin-top: 16px;
+        }
+    }
+
+    .btn-confirm {
+        position: absolute;
+        bottom: 0;
+        display: inline-block;
+        font-size: 16px;
+        padding: 10px;
+        width: 100px;
+        border: 1px solid #c2b9b9; /*no*/
+        border-radius: 4px;
+        margin-right: 30px;
+        margin-left: 30px;
+        cursor: pointer;
+        &:hover {
+            background-color: #55505046;
+        }
+    }
+    .btn-cancle{
+        position: absolute;
+        bottom: 0;
+        left: 150px;
+        display: inline-block;
+        font-size: 16px;
+        padding: 10px;
+        width: 100px;
+        border: 1px solid #c2b9b9; /*no*/
+        border-radius: 4px;
+        margin-right: 30px;
+        margin-left: 30px;
+        cursor: pointer;
+        &:hover {
+            background-color: #55505046;
+        }
+    }
+
+    }
+
+}
